+++ asset_project/asset_project.Shared/Entities/WorkOrder.cs
@@ -2,11 +2,11 @@
 {
     public class WorkOrder
     {
-        public long Id { get; set; }
+        public int Id { get; set; }
 
         public Asset? Asset { get; set; }
-
-        public long AssetId { get; set; }
+        
+        public int AssetId { get; set; }
 
         public StatusType? StatusType { get; set; }
 
@@ -14,10 +14,10 @@
 
         public MaintenanceRequest? MaintenanceRequest { get; set; }
 
-        public long MaintenanceRequestId { get; set; }
+        public int MaintenanceRequestId { get; set; }
 
         public string? TechnicianAssigned { get; set; }
 
-        public long TechnicianId { get; set;}
+        public int TechnicianId { get; set;}
     }
 }
